feat: get templates from remote archive
